Modularize pathway expression overlay
